webui/requirements-ui.txt
loguru
opencv-python-headless
python-dotenv
gradio==3.35.2
gradio_client==0.2.9
## SEEM
Pillow==9.5.0 # DETECTRON 2 Requirement
pyyaml
json_tricks
yacs
scikit-learn
pandas
timm==0.4.12
numpy==1.23.5
einops
fvcore
sentencepiece
ftfy
regex
nltk
vision-datasets==0.2.2
diffdist
pyarrow
cityscapesscripts
shapely
scikit-image
mup
scann
kornia==0.6.4
torchmetrics==0.6.0
invisible-watermark
diffusers==0.11.1
accelerate
altair
openai
git+https://github.com/arogozhnikov/einops.git
git+https://github.com/MaureenZOU/detectron2-xyz.git
git+https://github.com/openai/whisper.git
